+++ packages/yew/src/dom_bundle/tests/layout_tests.rs
@@ -1,6 +1,7 @@
-use crate::html::{AnyScope, Scope};
+use crate::dom_bundle::{BNode, DomBundle, Reconcilable};
+use crate::html::AnyScope;
 use crate::scheduler;
-use crate::virtual_dom::{VDiff, VNode, VText};
+use crate::virtual_dom::VNode;
 use crate::{Component, Context, Html};
 use gloo::console::log;
 use web_sys::Node;
@@ -37,21 +38,20 @@ pub struct TestLayout<'a> {
 
 pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
     let document = gloo_utils::document();
-    let parent_scope: AnyScope = Scope::<Comp>::new(None).into();
+    let parent_scope: AnyScope = AnyScope::test();
     let parent_element = document.create_element("div").unwrap();
     let parent_node: Node = parent_element.clone().into();
     let end_node = document.create_text_node("END");
     parent_node.append_child(&end_node).unwrap();
-    let mut empty_node: VNode = VText::new("").into();
 
     // Tests each layout independently
     let next_sibling = NodeRef::new(end_node.into());
     for layout in layouts.iter() {
         // Apply the layout
-        let mut node = layout.node.clone();
+        let vnode = layout.node.clone();
         log!("Independently apply layout '{}'", layout.name);
 
-        node.apply(&parent_scope, &parent_element, next_sibling.clone(), None);
+        let (_, mut bundle) = vnode.attach(&parent_scope, &parent_element, next_sibling.clone());
         scheduler::start_now();
         assert_eq!(
             parent_element.inner_html(),
@@ -61,15 +61,15 @@ pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
         );
 
         // Diff with no changes
-        let mut node_clone = layout.node.clone();
+        let vnode = layout.node.clone();
 
         log!("Independently reapply layout '{}'", layout.name);
 
-        node_clone.apply(
+        vnode.reconcile_node(
             &parent_scope,
             &parent_element,
             next_sibling.clone(),
-            Some(node),
+            &mut bundle,
         );
         scheduler::start_now();
         assert_eq!(
@@ -80,12 +80,7 @@ pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
         );
 
         // Detach
-        empty_node.clone().apply(
-            &parent_scope,
-            &parent_element,
-            next_sibling.clone(),
-            Some(node_clone),
-        );
+        bundle.detach(&parent_element, false);
         scheduler::start_now();
         assert_eq!(
             parent_element.inner_html(),
@@ -96,16 +91,16 @@ pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
     }
 
     // Sequentially apply each layout
-    let mut ancestor: Option<VNode> = None;
+    let mut bundle: Option<BNode> = None;
     for layout in layouts.iter() {
-        let mut next_node = layout.node.clone();
+        let next_vnode = layout.node.clone();
 
         log!("Sequentially apply layout '{}'", layout.name);
-        next_node.apply(
+        next_vnode.reconcile_sequentially(
             &parent_scope,
             &parent_element,
             next_sibling.clone(),
-            ancestor,
+            &mut bundle,
         );
         scheduler::start_now();
         assert_eq!(
@@ -114,19 +109,18 @@ pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
             "Sequential apply failed for layout '{}'",
             layout.name,
         );
-        ancestor = Some(next_node);
     }
 
     // Sequentially detach each layout
     for layout in layouts.into_iter().rev() {
-        let mut next_node = layout.node.clone();
+        let next_vnode = layout.node.clone();
 
         log!("Sequentially detach layout '{}'", layout.name);
-        next_node.apply(
+        next_vnode.reconcile_sequentially(
             &parent_scope,
             &parent_element,
             next_sibling.clone(),
-            ancestor,
+            &mut bundle,
         );
         scheduler::start_now();
         assert_eq!(
@@ -135,11 +129,12 @@ pub fn diff_layouts(layouts: Vec<TestLayout<'_>>) {
             "Sequential detach failed for layout '{}'",
             layout.name,
         );
-        ancestor = Some(next_node);
     }
 
     // Detach last layout
-    empty_node.apply(&parent_scope, &parent_element, next_sibling, ancestor);
+    if let Some(bundle) = bundle {
+        bundle.detach(&parent_element, false);
+    }
     scheduler::start_now();
     assert_eq!(
         parent_element.inner_html(),
