--- packages/yew/src/tests/mod.rs
@@ -1,1 +1,0 @@
-pub mod layout_tests;
