+++ packages/yew/src/virtual_dom/vnode.rs
@@ -1,14 +1,11 @@
 //! This module contains the implementation of abstract virtual node.
 
-use super::{Key, VChild, VComp, VDiff, VList, VPortal, VSuspense, VTag, VText};
-use crate::html::{AnyScope, BaseComponent, NodeRef};
-use gloo::console;
+use super::{Key, VChild, VComp, VList, VPortal, VSuspense, VTag, VText};
+use crate::html::BaseComponent;
 use std::cmp::PartialEq;
 use std::fmt;
 use std::iter::FromIterator;
-use wasm_bindgen::JsCast;
-
-use web_sys::{Element, Node};
+use web_sys::Node;
 
 /// Bind virtual element to a DOM reference.
 #[derive(Clone)]
@@ -30,177 +27,21 @@ pub enum VNode {
 }
 
 impl VNode {
-    pub fn key(&self) -> Option<Key> {
+    pub fn key(&self) -> Option<&Key> {
         match self {
-            VNode::VComp(vcomp) => vcomp.key.clone(),
-            VNode::VList(vlist) => vlist.key.clone(),
+            VNode::VComp(vcomp) => vcomp.key.as_ref(),
+            VNode::VList(vlist) => vlist.key.as_ref(),
             VNode::VRef(_) => None,
-            VNode::VTag(vtag) => vtag.key.clone(),
+            VNode::VTag(vtag) => vtag.key.as_ref(),
             VNode::VText(_) => None,
             VNode::VPortal(vportal) => vportal.node.key(),
-            VNode::VSuspense(vsuspense) => vsuspense.key.clone(),
+            VNode::VSuspense(vsuspense) => vsuspense.key.as_ref(),
         }
     }
 
-    /// Returns true if the [VNode] has a key without needlessly cloning the key.
+    /// Returns true if the [VNode] has a key.
     pub fn has_key(&self) -> bool {
-        match self {
-            VNode::VComp(vcomp) => vcomp.key.is_some(),
-            VNode::VList(vlist) => vlist.key.is_some(),
-            VNode::VRef(_) | VNode::VText(_) => false,
-            VNode::VTag(vtag) => vtag.key.is_some(),
-            VNode::VPortal(vportal) => vportal.node.has_key(),
-            VNode::VSuspense(vsuspense) => vsuspense.key.is_some(),
-        }
-    }
-
-    /// Returns the first DOM node if available
-    pub(crate) fn first_node(&self) -> Option<Node> {
-        match self {
-            VNode::VTag(vtag) => vtag.reference().cloned().map(JsCast::unchecked_into),
-            VNode::VText(vtext) => vtext
-                .reference
-                .as_ref()
-                .cloned()
-                .map(JsCast::unchecked_into),
-            VNode::VComp(vcomp) => vcomp.node_ref.get(),
-            VNode::VList(vlist) => vlist.get(0).and_then(VNode::first_node),
-            VNode::VRef(node) => Some(node.clone()),
-            VNode::VPortal(vportal) => vportal.next_sibling(),
-            VNode::VSuspense(vsuspense) => vsuspense.first_node(),
-        }
-    }
-
-    /// Returns the first DOM node that is used to designate the position of the virtual DOM node.
-    pub(crate) fn unchecked_first_node(&self) -> Node {
-        match self {
-            VNode::VTag(vtag) => vtag
-                .reference()
-                .expect("VTag is not mounted")
-                .clone()
-                .into(),
-            VNode::VText(vtext) => {
-                let text_node = vtext.reference.as_ref().expect("VText is not mounted");
-                text_node.clone().into()
-            }
-            VNode::VComp(vcomp) => vcomp.node_ref.get().unwrap_or_else(|| {
-                #[cfg(not(debug_assertions))]
-                panic!("no node_ref; VComp should be mounted");
-
-                #[cfg(debug_assertions)]
-                panic!(
-                    "no node_ref; VComp should be mounted after: {:?}",
-                    crate::virtual_dom::vcomp::get_event_log(vcomp.id),
-                );
-            }),
-            VNode::VList(vlist) => vlist
-                .get(0)
-                .expect("VList is not mounted")
-                .unchecked_first_node(),
-            VNode::VRef(node) => node.clone(),
-            VNode::VPortal(_) => panic!("portals have no first node, they are empty inside"),
-            VNode::VSuspense(vsuspense) => {
-                vsuspense.first_node().expect("VSuspense is not mounted")
-            }
-        }
-    }
-
-    pub(crate) fn move_before(&self, parent: &Element, next_sibling: &Option<Node>) {
-        match self {
-            VNode::VList(vlist) => {
-                for node in vlist.iter() {
-                    node.move_before(parent, next_sibling);
-                }
-            }
-            VNode::VComp(vcomp) => {
-                vcomp
-                    .root_vnode()
-                    .expect("VComp has no root vnode")
-                    .move_before(parent, next_sibling);
-            }
-            VNode::VPortal(_) => {} // no need to move portals
-            _ => super::insert_node(&self.unchecked_first_node(), parent, next_sibling.as_ref()),
-        };
-    }
-}
-
-impl VDiff for VNode {
-    /// Remove VNode from parent.
-    fn detach(&mut self, parent: &Element, parent_to_detach: bool) {
-        match *self {
-            VNode::VTag(ref mut vtag) => vtag.detach(parent, parent_to_detach),
-            VNode::VText(ref mut vtext) => vtext.detach(parent, parent_to_detach),
-            VNode::VComp(ref mut vcomp) => vcomp.detach(parent, parent_to_detach),
-            VNode::VList(ref mut vlist) => vlist.detach(parent, parent_to_detach),
-            VNode::VRef(ref node) => {
-                if parent.remove_child(node).is_err() {
-                    console::warn!("Node not found to remove VRef");
-                }
-            }
-            VNode::VPortal(ref mut vportal) => vportal.detach(parent, parent_to_detach),
-            VNode::VSuspense(ref mut vsuspense) => vsuspense.detach(parent, parent_to_detach),
-        }
-    }
-
-    fn shift(&self, previous_parent: &Element, next_parent: &Element, next_sibling: NodeRef) {
-        match *self {
-            VNode::VTag(ref vtag) => vtag.shift(previous_parent, next_parent, next_sibling),
-            VNode::VText(ref vtext) => vtext.shift(previous_parent, next_parent, next_sibling),
-            VNode::VComp(ref vcomp) => vcomp.shift(previous_parent, next_parent, next_sibling),
-            VNode::VList(ref vlist) => vlist.shift(previous_parent, next_parent, next_sibling),
-            VNode::VRef(ref node) => {
-                previous_parent.remove_child(node).unwrap();
-                next_parent
-                    .insert_before(node, next_sibling.get().as_ref())
-                    .unwrap();
-            }
-            VNode::VPortal(ref vportal) => {
-                vportal.shift(previous_parent, next_parent, next_sibling)
-            }
-            VNode::VSuspense(ref vsuspense) => {
-                vsuspense.shift(previous_parent, next_parent, next_sibling)
-            }
-        }
-    }
-
-    fn apply(
-        &mut self,
-        parent_scope: &AnyScope,
-        parent: &Element,
-        next_sibling: NodeRef,
-        ancestor: Option<VNode>,
-    ) -> NodeRef {
-        match *self {
-            VNode::VTag(ref mut vtag) => vtag.apply(parent_scope, parent, next_sibling, ancestor),
-            VNode::VText(ref mut vtext) => {
-                vtext.apply(parent_scope, parent, next_sibling, ancestor)
-            }
-            VNode::VComp(ref mut vcomp) => {
-                vcomp.apply(parent_scope, parent, next_sibling, ancestor)
-            }
-            VNode::VList(ref mut vlist) => {
-                vlist.apply(parent_scope, parent, next_sibling, ancestor)
-            }
-            VNode::VRef(ref mut node) => {
-                if let Some(mut ancestor) = ancestor {
-                    // We always remove VRef in case it's meant to be used somewhere else.
-                    if let VNode::VRef(n) = &ancestor {
-                        if node == n {
-                            return NodeRef::new(node.clone());
-                        }
-                    }
-                    ancestor.detach(parent, false);
-                }
-                super::insert_node(node, parent, next_sibling.get().as_ref());
-                NodeRef::new(node.clone())
-            }
-            VNode::VPortal(ref mut vportal) => {
-                vportal.apply(parent_scope, parent, next_sibling, ancestor)
-            }
-            VNode::VSuspense(ref mut vsuspense) => {
-                vsuspense.apply(parent_scope, parent, next_sibling, ancestor)
-            }
-        }
+        self.key().is_some()
     }
 }
 
@@ -242,6 +83,13 @@ impl From<VSuspense> for VNode {
     #[inline]
     fn from(vsuspense: VSuspense) -> Self {
         VNode::VSuspense(vsuspense)
+    }
+}
+
+impl From<VPortal> for VNode {
+    #[inline]
+    fn from(vportal: VPortal) -> Self {
+        VNode::VPortal(vportal)
     }
 }
 
@@ -299,9 +147,9 @@ impl PartialEq for VNode {
 
 #[cfg(feature = "ssr")]
 mod feat_ssr {
-    use futures::future::{FutureExt, LocalBoxFuture};
-
     use super::*;
+    use crate::html::AnyScope;
+    use futures::future::{FutureExt, LocalBoxFuture};
 
     impl VNode {
         // Boxing is needed here, due to: https://rust-lang.github.io/async-book/07_workarounds/04_recursion.html
@@ -333,38 +181,5 @@ mod feat_ssr {
             }
             .boxed_local()
         }
-    }
-}
-
-#[cfg(test)]
-mod layout_tests {
-    use super::*;
-    use crate::tests::layout_tests::{diff_layouts, TestLayout};
-
-    #[cfg(feature = "wasm_test")]
-    use wasm_bindgen_test::{wasm_bindgen_test as test, wasm_bindgen_test_configure};
-
-    #[cfg(feature = "wasm_test")]
-    wasm_bindgen_test_configure!(run_in_browser);
-
-    #[test]
-    fn diff() {
-        let document = gloo_utils::document();
-        let vref_node_1 = VNode::VRef(document.create_element("i").unwrap().into());
-        let vref_node_2 = VNode::VRef(document.create_element("b").unwrap().into());
-
-        let layout1 = TestLayout {
-            name: "1",
-            node: vref_node_1,
-            expected: "<i></i>",
-        };
-
-        let layout2 = TestLayout {
-            name: "2",
-            node: vref_node_2,
-            expected: "<b></b>",
-        };
-
-        diff_layouts(vec![layout1, layout2]);
     }
 }
