+++ tools/process-benchmark-results/Makefile.toml
@@ -1,0 +1,5 @@
+# Extends the root Makefile.toml
+
+[tasks.doc-test]
+command = "echo"
+args = ["No doctests run for process-benchmark-results"]
